Cleanup Commit.  I'm removing MORE dependencies, improved_inlines.  You might not want model based inlines, and that's cool.  Some initial fixes to the README, but that is going to require more attention, as are the docs in general.
The example project is now a single directory, contrib is totally gone, contrib.images is now example_project.sample_image_insert

I'm removing tinymce and adding it as a git submodule.  This complicates development a little, but makes the repository... and my management of it a little lighter.

Requirements file is _ridiculously_ lighter.